Implement different named PoVNETS
This allows the creation of multiple independent PoVNETs. Each one is
identified by a name which is hardened into a "PoVNET genesis" block,
which is automatically positioned at height 1. Validation rules will
ensure that a node from PoVNET=test1 never be able to accept blocks
from PoVNET=test2. This is done by checking the expected PoVNET genesis
block.

The genesis block of the PoVNET is the same as the one from regtest. This
starts the PoVNET with a very low difficulty, allowing us to fill up
needed balances for masternodes very fast.

Also, the PoVNET name is put into the sub-version of the VERSION message.
If a node connects to the wrong network, it will immediately be disconnected.